+++ pyccl/background.py
@@ -16,6 +16,7 @@ import numpy as np
 from . import ccllib as lib
 from .pyutils import _vectorize_fn, _vectorize_fn3
 from .pyutils import _vectorize_fn4, _vectorize_fn5
+from .parameters import physical_constants
 
 species_types = {
     'critical': lib.species_crit_label,
@@ -291,7 +292,6 @@ def sigma_critical(cosmo, a_lens, a_source):
         float or array_like: :math:`\\Sigma_{\\mathrm{crit}}` in units
         of :math:`\\M_{\\odot}/Mpc^2`
     """
-    physical_constants = lib.cvar.constants
     Ds = angular_diameter_distance(cosmo, a_source, a2=None)
     Dl = angular_diameter_distance(cosmo, a_lens, a2=None)
     Dls = angular_diameter_distance(cosmo, a_lens, a_source)
