+++ pyccl/halos/halo_model.py
@@ -12,9 +12,8 @@ from ..background import rho_x
 from ..pyutils import _spline_integrate
 from .. import background
 from ..errors import CCLWarning
+from ..parameters import physical_constants
 import numpy as np
-
-physical_constants = lib.cvar.constants
 
 
 class HMCalculator(object):
